% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untheme.R
\name{create_field_set}
\alias{create_field_set}
\title{Create a field set with a given icon, label, and selectInput}
\usage{
create_field_set(
  icon_name,
  label_text,
  input_id,
  input_choices,
  input_selected,
  numeric_input = FALSE
)
}
\arguments{
\item{icon_name}{Name of the icon used next to the label.}

\item{label_text}{Text for the label describing the input field.}

\item{input_id}{ID for the select or numeric input, used for input retrieval in server logic.}

\item{input_choices}{Choices for the select input (ignored if numeric_input is TRUE).}

\item{input_selected}{Default selected choice for the select input or default value for the numeric input.}

\item{numeric_input}{Logical indicating whether to use a numeric input instead of a select input (default is FALSE).}
}
\description{
This function creates a field set for the Shiny application, consisting of an icon, a label, and a select or numeric input.
The function allows for easy generation of a consistent UI element for different types of inputs.
}
