% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untheme.R
\name{detect_font_size}
\alias{detect_font_size}
\title{Detect Appropriate Font Size Based on Screen Width}
\usage{
detect_font_size(screen_width)
}
\arguments{
\item{screen_width}{Numeric value representing the width of the screen in pixels.}
}
\value{
A list containing three elements: \code{title} (the font size for titles),
\code{font} (the font size for general text), and \code{type} (a string indicating the
device type: "screen", "tablet", or "mobile").
}
\description{
This function determines the appropriate font sizes for titles and general text
based on the provided screen width. It categorizes devices into three types:
screen (large screens), tablet, and mobile. It returns a list containing the
title font size, general font size, and the device type.
}
