% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untheme.R
\name{generate_ids}
\alias{generate_ids}
\title{Generate Shiny Element IDs}
\usage{
generate_ids(id)
}
\arguments{
\item{id}{A character string representing the base ID.}
}
\value{
A named list containing three elements: plot_id, download_plot_id, and download_data_id.
}
\description{
Generate Shiny Element IDs
}
