% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untheme.R
\name{plots_tabset}
\alias{plots_tabset}
\title{Generate the plot modules for an arbitrary number of plots.}
\usage{
plots_tabset(input, output, id, plot_object)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{id}{Unique identifier for the plot module.}

\item{plot_object}{An object containing a reactive expression for the plot and its filename.}
}
\description{
Generate the plot modules for an arbitrary number of plots.
}
