% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untheme.R
\name{sidebar_layout_responsive}
\alias{sidebar_layout_responsive}
\title{Create a custom sidebar layout, mobile responsive}
\usage{
sidebar_layout_responsive(sidebar, main_panel)
}
\arguments{
\item{sidebar}{A list with UI components for the sidebar}

\item{main_panel}{A div with UI components for the main panel}
}
\description{
This custom sidebar panel is responsive: when below 1400px the sidebar
will be moved to be on top of the plot.
}
