% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untheme.R
\name{fluidUnTheme}
\alias{fluidUnTheme}
\title{Create a customized Shiny UI with additional CSS and HTML}
\usage{
fluidUnTheme(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[shiny]{div}} function.}
}
\value{
A \code{shiny.semantic::semanticPage} object with custom styling.
}
\description{
Create a customized Shiny UI with additional CSS and HTML
}
