% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untheme.R
\name{plots_tabset}
\alias{plots_tabset}
\title{Generate the plot modules for an arbitrary number of plots.}
\usage{
plots_tabset(...)
}
\arguments{
\item{...}{A variable number of reactive expressions, each returning a ggplot2 object for rendering.}
}
\description{
This function facilitates the generation of multiple plot modules in a Shiny application, particularly following a simulation or data analysis process.
It accepts a variable number of reactive expressions, each returning a ggplot2 object, and renders these plots within the application.
}
