% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untheme.R
\name{plotWithDownloadButtons}
\alias{plotWithDownloadButtons}
\title{Create a server component for rendering and downloading a plot}
\usage{
plotWithDownloadButtons(
  input,
  output,
  session,
  ggplot_obj,
  file_name,
  update_ggplot_func = NULL
)
}
\arguments{
\item{input}{A list of input values from the Shiny UI.}

\item{output}{A list of output values to be modified by this function.}

\item{session}{The Shiny session object.}

\item{ggplot_obj}{A reactive expression returning a list with a named ggplot object (gg) and a named plotly plot (plotly).}

\item{file_name}{The file name to use for the plot download and the data.}

\item{update_ggplot_func}{An optional function for updating the ggplot object
based on user input.}
}
\value{
None. This function modifies the \code{output} list in-place.
}
\description{
This function creates a Shiny server component for rendering a ggplot object
and providing download buttons for the plot and data.
}
\seealso{
\code{\link[shiny]{reactive}}, \code{\link[shiny]{renderPlot}}, \code{\link[shiny]{downloadHandler}}
}
