% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untheme.R
\name{plotWithDownloadButtonsUI}
\alias{plotWithDownloadButtonsUI}
\title{Create a UI component with a plot and optional download buttons}
\usage{
plotWithDownloadButtonsUI(id, radio_choices = NULL)
}
\arguments{
\item{id}{A unique identifier for the UI component.}

\item{radio_choices}{A vector of options for the radio buttons. If NULL, no radio
buttons will be displayed.}
}
\value{
A \code{shiny::sidebarLayout} object containing the plot and download buttons.
}
\description{
This function creates a Shiny UI component that includes a plot, a download button
for the plot, and optionally, a download button for the data and radio buttons for
selecting the scale type.
}
\examples{
\dontrun{
ui <- plotWithDownloadButtonsUI("plot1", radio_choices = c("Absolute", "Percentage"))
}
}
\seealso{
\code{\link[shiny]{NS}}, \code{\link[shiny]{downloadButton}}, \code{\link[shiny]{plotOutput}}
}
